+++ Backend/Functionality/Organising.py
@@ -1,4 +1,5 @@
 import logging
+import os
 
 from AiOrchestration.AiOrchestrator import AiOrchestrator
 from Data.CategoryManagement import CategoryManagement
@@ -6,7 +7,7 @@ from Data.Configuration import Configuration
 from Data.Files.StorageMethodology import StorageMethodology
 from Data.NodeDatabaseManagement import NodeDatabaseManagement as nodeDB
 from Data.UserContextManagement import UserContextManagement
-from Utilities.Contexts import set_functionality_context
+from Utilities.Contexts import set_functionality_context, get_category_context
 from Constants.Instructions import SUMMARISER_SYSTEM_INSTRUCTIONS
 from Utilities.Utility import Utility
 
@@ -26,16 +27,21 @@ class Organising:
         return file_references
 
     @staticmethod
-    def save_file(content: str, category: str, file_path: str, overwrite=True):
+    def save_file(content: str, category_id: str, filename: str, overwrite=True):
         """
         Stores the system in the database, both the content in storage and the file node representation
 
         :parameter content: As in file content
-        :parameter category: With respect to files the folder files are stored in
-        :parameter file_path: the full file address including folder directory and file name with extension.
+        :parameter category_id: The id of the category the file will belong to
+        :parameter filename: the filename of the file including extension, excluding category folder
+        :parameter overwrite: Whether any existing file should be overwritten with this files content
         """
-        nodeDB().create_file_node(category, file_path)
+        file_path = os.path.join(category_id, filename)
+
+        file_uuid = nodeDB().create_file_node(category_id, file_path)
         StorageMethodology().select().save_file(content, file_path, overwrite=overwrite)
+
+        return file_uuid
 
     @staticmethod
     def store_prompt_data(
@@ -52,7 +58,7 @@ class Organising:
         :return:
         """
         config = Configuration.load_config()
-        set_functionality_context(None)  # ToDo: Really different contexts should be isolated by thread.
+        set_functionality_context(None)  # ToDo: Really different functionalities should be isolated by thread.
 
         Utility.execute_with_retries(
             lambda: nodeDB().populate_user_prompt_node(category, user_prompt, response_message)
@@ -61,8 +67,6 @@ class Organising:
         if config['beta_features']['user_context_enabled']:
             terms = UserContextManagement.extract_terms_from_input([user_prompt])
             nodeDB().create_user_topic_nodes(terms)
-
-        CategoryManagement.stage_files(category)
 
     @staticmethod
     def summarise_content(content: str):
@@ -80,7 +84,7 @@ class Organising:
 
 if __name__ == '__main__':
     parent_class_name = "EncyclopediaManagementInterface.py"
-    child_class_name = "UserEncyclopediaMangement.py"
+    child_class_name = "UserEncyclopediaManagement.py"
     # Organising.get_relevant_files(
     #     "WIP",
     #     [f"Hey can you compare {parent_class_name} and the child class {child_class_name}?"]